backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
alembic
asyncpg
pydantic
pydantic-settings
python-dotenv
stripe
openai
anthropic
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
httpx